+++ pyjsonrpc/rpcrequest.py
@@ -2,50 +2,93 @@
 # coding: utf-8
 
 import sys
-import errors
 import traceback
 import uuid
-from jsontools import json, ParseError
+import rpcerror
+from rpcjson import json, JsonParseError
+
+
+class Request(dict):
+    """
+    JSON-RPC-Request
+    """
+
+    def __init__(
+        self,
+        jsonrpc = None,
+        method = None,
+        id = None,
+        params = None
+    ):
+        dict.__init__(
+            self,
+            jsonrpc = jsonrpc,
+            method = method,
+            id = id,
+            params = params
+        )
+
+        self.jsonrpc = jsonrpc
+        self.method = method
+        self.id = id
+        self.params = params
+
+
+    def get_splitted_params(self):
+        """
+        Split positional and named params
+
+        :returns: positional_params, named_params
+        """
+
+        positional_params = []
+        named_params = {}
+        params = self.get("params", [])
+        if isinstance(params, list):
+            positional_params = params
+        elif isinstance(params, dict):
+            positional_params = params.get("__args", [])
+            if positional_params:
+                del params["__args"]
+            named_params = params
+
+        return positional_params, named_params
 
 
 def parse_request_json(json_string):
     """
-    Returns RPC-request as dictionary or as list with dictionaries
-
-    :return: Dictionary or list with RPC-request-dictionaries.
-        Syntax::
-
-            {
-                "jsonrpc": "<json_rpc_version>",
-                "method": "<method_name>",
-                "id": "<id>",
-                "params": [<param>, ...]|{"<param_name>": <param_value>}
-            }
-            or
-            [
-                {
-                    "jsonrpc": "<json_rpc_version>",
-                    "method": "<method_name>",
-                    "id": "<id>",
-                    "params": [<param>, ...]|{"<param_name>": <param_value>}
-                },
-                ...
-            ]
+    Returns RPC-request as dictionary or as list with requests
     """
 
     # No JSON-String
     if json_string is None:
-        raise errors.InvalidRequest()
+        raise rpcerror.InvalidRequest()
 
     # Parse
     try:
         data = json.loads(json_string)
-    except ParseError:
+    except JsonParseError:
         traceback_info = "".join(traceback.format_exception(*sys.exc_info()))
-        raise errors.ParseError(data = traceback_info)
+        raise rpcerror.ParseError(data = traceback_info)
 
-    # Finished
-    return data
+    # Create request(s)
+    if isinstance(data, list):
+        requests = []
+        for item in data:
+            requests.append(Request(
+                jsonrpc = item.get("jsonrpc"),
+                method = str(item.get("method", "")),
+                id = item.get("id"),
+                params = item.get("params", [])
+            ))
+        return requests
+    else:
+        return Request(
+            jsonrpc = data.get("jsonrpc"),
+            method = str(data.get("method", "")),
+            id = data.get("id"),
+            params = data.get("params", [])
+        )
 
 
 def create_request_json(method, *args, **kwargs):
@@ -62,7 +105,7 @@ def create_request_json(method, *args, **kwargs):
     data = {
         "method": unicode(method),
         "id": unicode(uuid.uuid4()),
-        "jsonrpc": "2.0",
+        "jsonrpc": u"2.0",
         "params": params
     }
     return json.dumps(data)